# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Agent AI Platform API
  version: "1.0.0"
  description: >
    Modular Monolith · Hexagonal · Plugin · Event-Driven.
    Auth: Firebase ID Token (local JWT verify). Tenant (workspace_id) derived from identity, never client-supplied.
    Errors: RFC 9457 Problem Details. Pagination: cursor-based. JSON: snake_case.
    All collection endpoints are wrapped as { data: [...], meta: { next_cursor, limit } } (API-04);
    bounded lists set next_cursor: null. Every response returns X-Correlation-Id.
    The interactive WebSocket endpoint /api/v1/ws is out of scope for OpenAPI and is specified in 03-api-spec.md §3.2.
servers:
  - url: /api/v1
security:
  - firebaseBearer: []
...
paths:
...
  /agents:
    get:
      tags: [agents]
      operationId: listAgents
      responses:
        "200": { description: OK, headers: { X-Correlation-Id: { $ref: "#/components/headers/CorrelationId" } }, content: { application/json: { schema: { $ref: "#/components/schemas/AgentPage" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /agents/{agent_key}:
    get:
      tags: [agents]
      operationId: getAgent
      parameters: [ { $ref: "#/components/parameters/AgentKey" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AgentOut" } } } }
        default: { $ref: "#/components/responses/Problem" }
  /agents/{agent_key}/invoke:
    post:
      tags: [agents]
      operationId: invokeAgent
      description: "Run an agent. Set Accept: text/event-stream for SSE streaming, else JSON."
      parameters: [ { $ref: "#/components/parameters/AgentKey" }, { $ref: "#/components/parameters/CorrelationId" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AgentInvokeIn" } } } }
      responses:
        "200":
          description: Completion (JSON) or event stream (SSE)
          content:
            application/json: { schema: { $ref: "#/components/schemas/AgentInvokeOut" } }
            text/event-stream: { schema: { type: string, description: "SSE: token|tool_call|final|error events" } }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AgentKey: { name: agent_key, in: path, required: true, schema: { type: string } }
...
    CorrelationId: { name: X-Correlation-Id, in: header, required: false, schema: { type: string } }
...
  headers:
    CorrelationId:
      description: "Request correlation id, echoed on every response (generated if absent). See 03 §0."
      schema: { type: string }
  responses:
    Problem:
      description: RFC 9457 Problem Details
      headers:
        X-Correlation-Id: { $ref: "#/components/headers/CorrelationId" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, correlation_id]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: "stable machine code, e.g. files.too_large" }
        correlation_id: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    PageMeta:
      type: object
      properties: { next_cursor: { type: string, nullable: true, description: "opaque cursor; null on the last (or unbounded) page" }, limit: { type: integer } }
...
    AgentOut:
      type: object
      properties:
        key: {type: string}
        name: {type: string}
        version: {type: string}
        description: {type: string}
        capabilities: { type: array, items: {type: string} }
        required_permissions: { type: array, items: {type: string} }
...
    AgentInvokeIn:
      type: object
      required: [input]
      properties:
        conversation_id: { type: string, format: uuid, nullable: true }
        space_id:
          type: string
          format: uuid
          nullable: true
          description: "The space a NEW thread is opened in. Optional because a request that names a conversation_id inherits that thread's space and must not restate it; naming NEITHER is a 422."
        input: { type: object, additionalProperties: true }
        stream: { type: boolean, default: false }
    AgentInvokeOut:
      type: object
      properties:
        conversation_id: {type: string, format: uuid}
        message: { $ref: "#/components/schemas/MessageOut" }
        usage: { $ref: "#/components/schemas/Usage" }
    Usage:
      type: object
      properties: { prompt_tokens: {type: integer}, completion_tokens: {type: integer} }
...
    MessageOut:
      type: object
      properties:
        id: {type: string, format: uuid}
        role: {type: string, enum: [user, assistant, system, tool]}
        content: {type: object, additionalProperties: true}
        token_count: {type: integer, nullable: true}
        seq: {type: integer}
        created_at: {type: string, format: date-time, description: "UTC ISO-8601, trailing Z (DAT-06)", pattern: "Z$"}
...
    AgentPage:
      type: object
      properties: { data: { type: array, items: { $ref: "#/components/schemas/AgentOut" } }, meta: { $ref: "#/components/schemas/PageMeta" } }
...